Extract code to filter annotation tracks into new module
